requirements.txt
requests
customtkinter
pillow
timezonefinder